change the datatype of the variables declared in ExerciseOne.c from int to float

+++ ExerciseOne.c
@@ -7,7 +7,7 @@
 
 /*The first exercise will prompt the user to enter numbers one by one.*/
 /*When the user enters 0 or a negative number the program will display the largest number entered */
-/*and display the greatest as well as smallest integer*/
+
 
 #include<stdio.h>
 int main(){
@@ -16,8 +16,8 @@ int main(){
     /*The num variable will be used to store the greatest number the user inputs*/
     /*intialValue will be used to store the number the user inputs(not necessarily the largest number but whatever the user inputs)*/
 
-    int num = 0;
-    int intialValue;
+    float num = 0;
+    float intialValue;
 
     /*This for loop will only be displayed when the user inputs a number besides 0 or negative value*/
     /*As soon as we run the program this loop will be displayed and thats because intialValue has a value different from 0*/
@@ -25,7 +25,7 @@ int main(){
     
     for (int i = 0; intialValue > 0; i++){ 
         printf("Please enter the number: ");
-        scanf("%d", &intialValue);
+        scanf("%f", &intialValue);
         num = num>intialValue ? num:intialValue;
         }
 
@@ -33,7 +33,7 @@ int main(){
     /*This is only dispalyed when the user enters a number thats less than or equal to zero*/
     /*num contains the greatest number.*/
     
-    printf("The greatest number is: %d.\n",num);
-    
+    printf("The greatest number is: %f.\n",num);
+
      return 0;
 }
